fix msvc
