+++ src/gui/gui_graphics.hpp
@@ -397,9 +397,4 @@ namespace ui {
 	void show_main_menu(sys::state& state);
 	int32_t ui_width(sys::state const& state);
 	int32_t ui_height(sys::state const& state);
-
-	void fire_event(sys::state& state, const dcon::national_event_id event_id);
-	void fire_event(sys::state& state, const dcon::free_national_event_id event_id);
-	void fire_event(sys::state& state, const dcon::provincial_event_id event_id);
-	void fire_event(sys::state& state, const dcon::free_provincial_event_id event_id);
 }
